Reorganize k8s common
